adding localtoc html to fix sidebar(#132)
**Context:**
Right-side navigation bar originally it showed the list of classes in
the same module, and now it shows the list of methods and attributes of
the current class.

**Description of the Change:**
Removing localtoc.html sidebar if statement so that it always references
the main toctree

**Benefits:**
Fixes issue with right-side navigation bar.

**Possible Drawbacks:**
May cause other unknown issues

**Related GitHub Issues:**